Finished the SearchGroups API Added a NotSet = -1 enum value to all API responses, which allows us to get rid of a nullable on the client side.
