+++ church.ccv.CCVRest/MobileApp/Util/MobileAppService.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using church.ccv.Actions;
 using church.ccv.CCVRest.Common;
 using church.ccv.CCVRest.MobileApp.Model;
+using church.ccv.Datamart.Model;
 using Rock;
 using Rock.Communication;
 using Rock.Data;
@@ -368,7 +369,7 @@ namespace church.ccv.CCVRest.MobileApp
 
             //const int InteractionComponent_Id = 156245; //Production Value - Can't use in MA3 until we refresh the server
             const int InteractionComponent_Id = 151588; //Temp MA3 value - Remove once we refresh MA3.
-            
+
             // first, see if there's already a device type matching this requester - We store device types 
             // with our defined InteractionDeviceTypeData above, and the userAgent string. That's it.
             InteractionDeviceTypeService interactionDeviceTypeService = new InteractionDeviceTypeService( rockContext );
@@ -420,6 +421,204 @@ namespace church.ccv.CCVRest.MobileApp
                 InteractionSessionId = interactionSession.Id
             };
             interactionService.Add( thisInteraction );
+        }
+
+        public static List<MobileAppGroupModel> GetMobileAppGroups( string nameKeyword,
+                                                                    string descriptionKeyword,
+                                                                    Location locationForDistance,
+                                                                    bool? requiresChildcare,
+                                                                    int? skip,
+                                                                    int top )
+        {
+            // Gets Neighborhood Groups, searches by the provided arguments, and returns matching values as MobileAppGroupModels
+
+            const int GroupTypeId_NeighborhoodGroup = 49;
+            const int GroupRoleId_NeighborhoodGroupCoach = 50;
+
+            // The id for the group description on Neighborhood groups. Used for joining the attributeValue if a descriptionKeyword is provided.
+            const int AttributeId_GroupDescription = 13055;
+
+            // The id for the group filters on Neighborhood groups. Used for joininig the attributeValue to see if it contains "Childcare"
+            const int AttributeId_GroupFilters = 42850;
+
+            const string GroupDescription_Key = "GroupDescription";
+            const string ChildcareDescription_Key = "Childcare";
+            const string FamilyPicture_Key = "FamilyPicture";
+
+            const string GroupFilters_Key = "GroupFilters";
+            const string ChildcareProvided_FilterKey = "Childcare Provided";
+
+            // First get all neighborhood groups, filtered by name and description if the caller provided those keywords
+            RockContext rockContext = new RockContext();
+
+            // get all groups of this group type that are public, and have a long/lat we can use
+            GroupService groupService = new GroupService( rockContext );
+            IEnumerable<Group> groupList = groupService.Queryable( "Schedule,GroupLocations.Location" ).AsNoTracking()
+                                                       .Where( a => a.GroupTypeId == GroupTypeId_NeighborhoodGroup && a.IsPublic == true )
+                                                       .Include( a => a.GroupLocations ).Where( a => a.GroupLocations.Any( x => x.Location.GeoPoint != null ) );
+
+            // if they provided name keywords, filter by those
+            if ( string.IsNullOrWhiteSpace( nameKeyword ) == false )
+            {
+                groupList = groupList.Where( a => a.Name.ToLower().Contains( nameKeyword.ToLower() ) );
+            }
+
+            // if they provided description, we need to join the attribute value table
+            if ( string.IsNullOrWhiteSpace( descriptionKeyword ) == false )
+            {
+                // Join the attribute value that defines the GroupDescription with the group
+                var avQuery = new AttributeValueService( rockContext ).Queryable().AsNoTracking().Where( av => av.AttributeId == AttributeId_GroupDescription );
+                var joinedQuery = groupList.Join( avQuery, g => g.Id, av => av.EntityId, ( g, av ) => new { Group = g, GroupDesc = av.Value } );
+
+                // see if the GroupDescription attribute value has the description keyword in it
+                groupList = joinedQuery.Where( g => g.GroupDesc.ToLower().Contains( descriptionKeyword.ToLower() ) ).Select( g => g.Group );
+            }
+
+            // if they require childcare, we again need to join the attribute value table
+            if ( requiresChildcare == true )
+            {
+                // Join the attribute vale that defines Group Filters (where Childcare is) with the group
+                var avQuery = new AttributeValueService( rockContext ).Queryable().AsNoTracking().Where( av => av.AttributeId == AttributeId_GroupFilters );
+                var joinedQuery = groupList.Join( avQuery, g => g.Id, av => av.EntityId, ( g, av ) => new { Group = g, Filter = av.Value } );
+
+                // see if the GroupFilter attribute value contains the ChildcareProvided filter key
+                groupList = joinedQuery.Where( g => g.Filter.Contains( ChildcareProvided_FilterKey ) ).Select( g => g.Group );
+            }
+
+
+            // calculate the distance of each of the group's locations from the specified geoFence
+            if ( locationForDistance != null )
+            {
+                // pull it into memory, because we have to in order to store distances on the location and sort by that
+                // (This could be optimized by creating a lookup table and sending that to sql)
+                groupList = groupList.ToList();
+
+                foreach ( var group in groupList )
+                {
+                    foreach ( var gl in group.GroupLocations )
+                    {
+                        // Calculate distance
+                        if ( gl.Location.GeoPoint != null )
+                        {
+                            double meters = gl.Location.GeoPoint.Distance( locationForDistance.GeoPoint ) ?? 0.0D;
+                            gl.Location.SetDistance( meters * Location.MilesPerMeter );
+                        }
+                    }
+                }
+
+                // and sort by the set distance
+                groupList = groupList.OrderBy( a => a.GroupLocations.First().Location.Distance ).ToList();
+            }
+
+
+            // grab the nth set
+            if ( skip.HasValue )
+            {
+                groupList = groupList.Skip( skip.Value ).ToList();
+            }
+
+            // and take the top amount
+            groupList = groupList.Take( top ).ToList();
+
+
+            // Now package the groups into GroupResult objects that store what the Mobile App cares about
+            var datamartPersonService = new DatamartPersonService( rockContext ).Queryable().AsNoTracking();
+            var personService = new PersonService( rockContext ).Queryable().AsNoTracking();
+            var binaryFileService = new BinaryFileService( rockContext ).Queryable().AsNoTracking();
+
+            List<MobileAppGroupModel> groupResultList = new List<MobileAppGroupModel>();
+
+            // now take only what we need from each group (drops our return package to about 2kb, from 40kb)
+            foreach ( Group group in groupList )
+            {
+                // now get the group leader. If there isn't one, we'll fail, because we don't want a group with no leader
+                GroupMember leader = group.Members.Where( gm => GroupRoleId_NeighborhoodGroupCoach == gm.GroupRole.Id ).SingleOrDefault();
+                if ( leader != null )
+                {
+                    // we are guaranteed that there will be a location object due to our initial query
+                    Location locationObj = group.GroupLocations.First().Location;
+
+                    MobileAppGroupModel groupResult = new MobileAppGroupModel()
+                    {
+                        Id = group.Id,
+
+                        Name = group.Name,
+                        Description = group.Description,
+
+                        Longitude = locationObj.Longitude.Value,
+                        Latitude = locationObj.Latitude.Value,
+                        DistanceFromSource = locationObj.Distance,
+
+                        MeetingTime = group.Schedule != null ? group.Schedule.FriendlyScheduleText : "",
+
+                        Street = locationObj.Street1,
+                        City = locationObj.City,
+                        State = locationObj.State,
+                        Zip = locationObj.PostalCode
+                    };
+
+                    // now find the leader in our datamart so that we can see who their Associate Pastor / Neighborhood Leader is
+                    var datamartPerson = datamartPersonService.Where( dp => dp.PersonId == leader.Person.Id ).SingleOrDefault();
+                    if ( datamartPerson != null )
+                    {
+                        groupResult.CoachName = leader.Person.NickName + " " + leader.Person.LastName;
+                        groupResult.CoachPhotoId = leader.Person.PhotoId;
+
+                        // if the leader has a neighborhood pastor (now called associate pastor) defined, take their values.
+                        if ( datamartPerson.NeighborhoodPastorId.HasValue )
+                        {
+                            // get the AP, but guard against a null value (could happen if the current ID is merged and the datamart hasn't re-run)
+                            Person associatePastor = personService.Where( p => p.Id == datamartPerson.NeighborhoodPastorId.Value ).SingleOrDefault();
+                            if ( associatePastor != null )
+                            {
+                                groupResult.AssociatePastorName = associatePastor.NickName + " " + associatePastor.LastName;
+                                groupResult.AssociatePastorPhotoId = associatePastor.PhotoId;
+                            }
+                        }
+                    }
+
+                    // Finally, load attributes so we can set additional group info
+                    group.LoadAttributes();
+
+                    if ( group.AttributeValues.ContainsKey( GroupDescription_Key ) )
+                    {
+                        groupResult.Description = group.AttributeValues[GroupDescription_Key].Value;
+                    }
+
+                    if ( group.AttributeValues.ContainsKey( FamilyPicture_Key ) )
+                    {
+                        Guid photoGuid = group.AttributeValues[FamilyPicture_Key].Value.AsGuid();
+
+                        // get the id so that we send consistent data down to the mobile app
+                        var photoObj = binaryFileService.Where( f => f.Guid == photoGuid ).SingleOrDefault();
+                        if ( photoObj != null )
+                        {
+                            groupResult.PhotoId = photoObj.Id;
+                        }
+                    }
+
+                    // get the childcare description whether the Childcare filter is set or NOT. This is
+                    // because some groups (like mine!) explain that they'd be willing to start childcare if the group grew.
+                    if ( group.AttributeValues.ContainsKey( ChildcareDescription_Key ) )
+                    {
+                        groupResult.ChildcareDesc = group.AttributeValues[ChildcareDescription_Key].Value;
+                    }
+
+                    // filters contain a comma delimited list of features the group offers. See if it has any.
+                    if ( group.AttributeValues.ContainsKey( GroupFilters_Key ) )
+                    {
+                        // The only one we currently care about it Childcare.
+                        if ( group.AttributeValues[GroupFilters_Key].Value.Contains( ChildcareProvided_FilterKey ) )
+                        {
+                            groupResult.Childcare = true;
+                        }
+                    }
+
+                    groupResultList.Add( groupResult );
+                }
+            }
+
+            return groupResultList;
         }
     }
 }
